Move the sheet handler into the new MainWindow struct.
